Update Rakudo Star version to 2019.03
This updates the Rakudo Star version to 2019.03 from 2018.10.

The maintainer for Rakudo Star changed between 2018.10 and 2019.03 so
the existing fingerprint used to retrieve the key and verify the tarball
is not valid for the new release. Updating to include the new
fingerprint along with the previous one; the old one will be used if the
user sets the "rakudo_version" arg to a release prior to 2019.03 when
building.

See [Github Issue #23](https://github.com/perl6/docker/issues/23)